pyproject.toml
[project]
name = "chief-of-staff"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.20",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
include = ["bot*", "mcp*"]
